requirements.txt
asgiref==3.8.1
Django==5.0.7
django-restframework==0.0.1
djangorestframework==3.15.2
psycopg2==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2